+++ eva/catalog/sql_config.py
@@ -12,6 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import os
 
 from sqlalchemy import create_engine, event
 from sqlalchemy.orm import scoped_session, sessionmaker
@@ -20,16 +21,18 @@ from eva.configuration.configuration_manager import ConfigurationManager
 
 IDENTIFIER_COLUMN = "_row_id"
 
-# import os
-# def prefix_worker_id(uri: str):
-#    try:
-#        worker_id = os.environ["PYTEST_XDIST_WORKER"]
-#        base = "eva_catalog.db"
-#        uri = uri.replace(base, str(worker_id) + "_" + base)
-#    except KeyError:
-#        # Single threaded mode
-#        pass
-#    return uri
+CATALOG_TABLES = [
+    "column_catalog",
+    "table_catalog",
+    "depend_column_and_udf_cache",
+    "udf_cache",
+    "udf_catalog",
+    "depend_udf_and_udf_cache",
+    "index_catalog",
+    "udfio_catalog",
+    "udf_cost_catalog",
+    "udf_metadata_catalog",
+]
 
 
 class SQLConfig:
@@ -58,10 +61,21 @@ class SQLConfig:
         Retrieves the database uri for connection from ConfigurationManager.
         """
         uri = ConfigurationManager().get_value("core", "catalog_database_uri")
-        # parallelize using xdist
-        # worker_uri = prefix_worker_id(str(uri))
+
+        # to parallelize tests using pytest-xdist
+        def prefix_worker_id_to_uri(uri: str):
+            try:
+                worker_id = os.environ["PYTEST_XDIST_WORKER"]
+                base = "eva_catalog.db"
+                # eva_catalog.db -> test_gw1_eva_catalog.db
+                uri = uri.replace(base, "test_" + str(worker_id) + "_" + base)
+            except KeyError:
+                pass
+            return uri
+
+        self.worker_uri = prefix_worker_id_to_uri(str(uri))
         # set echo=True to log SQL
-        self.engine = create_engine(uri)
+        self.engine = create_engine(self.worker_uri, isolation_level="SERIALIZABLE")
 
         if self.engine.url.get_backend_name() == "sqlite":
             # enforce foreign key constraint and wal logging for sqlite